refactor: rationalize filtering for datasource and indexed-data-store
allows to have a single business logic shared across the two datasources
